requirements.txt
streamlit>=1.28.0
rdkit-pypi>=2022.9.5
py3Dmol>=0.8.0
pandas>=2.0.3
numpy>=1.24.3
matplotlib>=3.7.1
seaborn>=0.12.2
torch>=2.0.1
scikit-learn>=1.3.0
requests>=2.31.0
Pillow>=9.0.0  # For image handling